+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * fix build setup when hosting elasticvue on a subdirectory
 * fix error when showing documents that include `id` prop, fixes [#219](https://github.com/cars10/elasticvue/issues/219)
 * save per_page & sort settings on indices page
+* show index creation date on indices page
 
 ## 1.0.4